feat(order) - 2차 리뷰 적용
https://chatgpt.com/share/678e3d9d-e790-8003-b7fc-b4750238b307

+++ order/src/main/java/com/smart/commerce/order/module/order/infrastructure/listener/event/OrderToPaymentEvent.java
@@ -3,6 +3,6 @@ package com.smart.commerce.order.module.order.infrastructure.listener.event;
 import org.jmolecules.event.types.DomainEvent;
 
 // TODO 필요한 값 설정 @희동
-public record OrderToPaymentEvent(int orderId, OrderStatus orderStatus) implements DomainEvent {
+public record OrderToPaymentEvent(Long orderId, OrderStatus orderStatus) implements DomainEvent {
 }
